Add button Dialog created. Some minior changes

+++ app/src/main/java/com/marcel/mycompany/screens/workers/Worker.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "workers_table")
-data class Worker(val name: String, val surName: String,val money:Float, var hours:Int=0) {
+data class Worker(val name: String, val surName: String,val money:Double, var hours:Double=0.0) {
     @PrimaryKey(autoGenerate = true)
     var worker_id:Int =0
 }
